+++ src/Renderer/StateAnimation.cpp
@@ -1,7 +1,7 @@
 #include "StateAnimation.h"
 
 namespace Renderer{
-	StateAnimation::StateAnimation(std::vector<std::pair<std::shared_ptr<Renderer::Sprite>, uint64_t>> frames, std::vector<std::string> sources, std::string nextState, bool uninterrupted)
+	StateAnimation::StateAnimation(std::vector<std::pair<std::shared_ptr<Renderer::Sprite>, double>> frames, std::vector<std::string> sources, std::string nextState, bool uninterrupted)
 		:m_frames(std::move(frames)),m_sources(std::move(sources)),m_nextState(nextState), m_uninterrupted(uninterrupted) {
 	}
